# docs/swagger.yaml
basePath: /
definitions:
  models.ChangeStatusRequest:
    properties:
      status:
        example: SUCCESS
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: incorrect user ID
        type: string
    type: object
  models.LogInRequest:
    properties:
      password:
        example: "123456789"
        type: string
      username:
        example: master_of_puppets
        type: string
    type: object
  models.LoginResponse:
    properties:
      token:
        type: string
    type: object
  models.NewPaymentRequest:
    properties:
      amount:
        example: 6535
        type: integer
      currency:
        example: RUB
        type: string
      user_email:
        example: secret@mail.ru
        type: string
      user_id:
        example: 5
        type: integer
    type: object
  models.NewPaymentResponse:
    properties:
      payment:
        $ref: '#/definitions/models.Payment'
    type: object
  models.Payment:
    properties:
      amount:
        example: 3456
        type: integer
      created_at:
        example: "2022-06-09T14:48:12.288326+03:00"
        type: string
      currency:
        example: MDA
        type: string
      id:
        example: 35
        type: integer
      status:
        example: NEW
        type: string
      updated_at:
        example: "2022-06-09T14:48:12.288326+03:00"
        type: string
      user_email:
        example: user@gmail.com
        type: string
      user_id:
        example: 31
        type: integer
    type: object
  models.PaymentStatusResponse:
    properties:
      status:
        example: ERROR
        type: string
    type: object
  models.PaymentsResponse:
    properties:
      payments:
        items:
          $ref: '#/definitions/models.Payment'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: Simple api for handling payments.
  title: Payment Service APi
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Use this route to get authorization token that will be needed to
        change payment's status( Add "Bearer + (token)" in ApiKeyAuth).
      parameters:
      - description: userdata
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LogInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: SignIn
      tags:
      - auth
  /payments/{id}/change:
    post:
      consumes:
      - application/json
      description: Allows to change status if current status is NEW on SUCCESS OR
        FAILURE. Authorize first!!!!!!!!!!!!!
      parameters:
      - description: Payment ID
        example: 50
        in: path
        name: id
        required: true
        type: integer
      - description: status name
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.ChangeStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Change payment's status
      tags:
      - Payments
  /payments/{id}/deny:
    delete:
      consumes:
      - application/json
      description: Deny payment if it exists and its status equals NEW or ERROR.
      parameters:
      - description: Payment ID
        example: 50
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Deny payment if it is possible
      tags:
      - Payments
  /payments/{id}/status:
    get:
      consumes:
      - application/json
      description: return payment's status if payment exists else returns error message.
      parameters:
      - description: Payment ID
        example: 50
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentStatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Returns payment's status
      tags:
      - Payments
  /payments/new:
    post:
      consumes:
      - application/json
      description: Create new payment from request's body.
      parameters:
      - description: payment info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.NewPaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NewPaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create new payment
      tags:
      - Payments
  /payments/user/{id}:
    get:
      consumes:
      - application/json
      description: Returns all payments with chosen user ID.
      parameters:
      - description: User ID
        example: 50
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Returns all payments with chosen user ID
      tags:
      - Payments
  /payments/user/email/{email}:
    get:
      consumes:
      - application/json
      description: Returns all payments with chosen user email.
      parameters:
      - description: User email
        example: user@gmail.com
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaymentsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Returns all payments with chosen user email
      tags:
      - Payments
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"